enables free database button during onboarding
