+++ configurator/frontend/src/ui/components/EmptyList/EmptyList.tsx
@@ -4,12 +4,9 @@ import { Modal } from 'antd';
 import { useHistory } from 'react-router-dom';
 // @Components
 import { EmptyListView } from './EmptyListView';
-// @Services
-import { useServices } from '@hooks/useServices';
-// @Utils
-import ApiKeyHelper from '@service/ApiKeyHelper';
-import { randomId } from '@util/numbers';
 import { reloadPage } from '@./lib/commons/utils';
+// @Commons
+import { createFreeDatabase } from '@./lib/commons/createFreeDatabase';
 
 export interface Props {
   title: ReactNode;
@@ -18,27 +15,10 @@ export interface Props {
 }
 
 const EmptyListComponent = ({ title, list, unit }: Props) => {
-  const services = useServices();
   const router = useHistory();
 
   const handleCreateFreeDatabase = useCallback<() => Promise<void>>(async() => {
-    const { destinations } = await services.initializeDefaultDestination();
-    services.analyticsService.track('create_database');
-    let helper = new ApiKeyHelper(services, { destinations });
-    await helper.init();
-    if (helper.keys.length === 0) {
-      const newKey: APIKey = {
-        uid: `${services.activeProject.id}.${randomId(5)}`,
-        serverAuth: `s2s.${services.activeProject.id}.${randomId(5)}`,
-        jsAuth: `js.${services.activeProject.id}.${randomId(5)}`,
-        origins: []
-      };
-      await services.storageService.save('api_keys', { keys: [newKey] }, services.activeProject.id);
-      helper = new ApiKeyHelper(services, { destinations, keys: [newKey] })
-    }
-    if (!helper.hasLinks()) {
-      await helper.link();
-    }
+    await createFreeDatabase();
     const modal = Modal.info({
       title: 'New destination has been created',
       content: <>
@@ -51,18 +31,13 @@ const EmptyListComponent = ({ title, list, unit }: Props) => {
       </>,
       onOk: () => reloadPage()
     });
-  }, [
-    services,
-    router,
-    services.activeProject.id
-  ]);
+  }, [router]);
 
   return (
     <EmptyListView
       title={title}
       list={list}
       unit={unit}
-      showFreeDatabaseSeparateButton={true}
       handleCreateFreeDatabase={handleCreateFreeDatabase}
     />
   );
